[Compose] avoid copying Compose-specific annotations to bridges
As a new feature since Kotlin 2.4, Kotlin compiler will copy
annotations from the bridge target methods to bridge methods. This may
affect Compose as `@FunctionKeyMeta` annotations are copied too, along
with the key that is supposed to be unique.
In order to fix that problem, the Compose-related annotations are added
to "ignored" list of annotations for such copying.