[FIR] Properly process `!is` branches in exhaustiveness checker
In the scope of KT-80461 the new simplified mechanism for checking
exhaustiveness with sealed subject was introduced. For `!is Type` it
considered all the sealed subclasses which are not subtype of `Type` to
be checked. But it's actually incorrect in case if the `Type` is an
inheritor of the sealed subclass itself, which caused false-negative
`NO_ELSE_IN_WHEN` happen.

^KT-82841 Fixed